refactor: clean auth module, remove token observer
